[integrations-api][test] Rename experimental tests to beta in dagster-airbyte
